v2.33.34 refactor (simplify PR-7): AddStop/ChangePoi 抽 shared poi-search helpers
/simplify finding: AddStopPage / ChangePoiPage 100% 複製 8 個
constants/functions (REGION_OPTIONS / CATEGORY_TABS / matchCategory /
normalizeSearchResults / poiTone / poiMeta / PoiCardTone / Tab)。Drift
存在: ChangePoi normalizeSearchResults cast-only 無 type check。

Extract src/lib/poiSearchHelpers.ts (取嚴格版 canonical):
- REGION_OPTIONS + RegionOption
- CATEGORY_TABS + PoiSearchCategory
- PoiCardTone + PoiSearchTab
- matchCategory + normalizeSearchResults + poiTone + poiMeta

兩個 page 改 import from shared lib，刪 local definitions。~150 LOC ↓

Full <PoiSearchTab> / <PoiFavoritesTab> component extraction 留 future PR。

Test source-grep 同步指向 poiSearchHelpers.ts。

270 files / 2092 tests pass.
